lab2 radi

+++ Lab02/main2.cpp
@@ -4,39 +4,69 @@
 #include <signal.h>
 #include <sys/types.h>
 #include <sys/wait.h>
-#include <string.h>
+#include <string>
+#include <sstream>
 #include <termios.h>
 #include <iostream>
 #include <list>
+#include <string.h>
 
 #define MAXARGS 10
 
 class Proces
 {
+
 public:
 	pid_t pid;
 	std::string name;
-	Proces(pid_t pid, std::string naziv) : pid(pid), name(name) {}
+	Proces(pid_t pid_c, std::string naziv_c) : pid(pid_c), name(naziv_c) {}
 };
 
 struct sigaction prije;
+std::list<Proces> still_running;
+
+bool is_running(pid_t pid)
+{
+	for (const Proces &proces : still_running)
+	{
+		if (proces.pid == pid)
+			return true;
+	}
+	return false;
+}
+
+void ukloni_proces_iz_running(pid_t pid_zavrsio)
+{
+	for (auto it = still_running.begin(); it != still_running.end(); ++it)
+	{
+		if (it->pid == pid_zavrsio)
+		{
+			still_running.erase(it);
+			std::cout << "Proces with PID: " << it->pid << " ended." << std::endl;
+			break;
+		}
+	}
+}
 
 void obradi_signal_zavrsio_neki_proces_dijete(int id)
 {
 	pid_t pid_zavrsio = waitpid(-1, NULL, WNOHANG);
-	if(pid_zavrsio > 0)
-		if(kill(pid_zavrsio, 0) == -1)
+	if (pid_zavrsio > 0)
+		if (kill(pid_zavrsio, 0) == -1)
+		{
+			ukloni_proces_iz_running(pid_zavrsio);
 			printf("\n[roditelj %d - SIGCHLD + waitpid] dijete %d zavrsilo s radom\n", (int)getpid(), pid_zavrsio);
+		}
 }
 
 pid_t pokreni_program(char *naredba[], int u_pozadini)
 {
 	pid_t pid_novi;
-	if((pid_novi = fork()) == 0)
+	if ((pid_novi = fork()) == 0)
 	{
 		sigaction(SIGINT, &prije, NULL);
 		setpgid(pid_novi, pid_novi);
-		if(!u_pozadini)
+		if (!u_pozadini)
 			tcsetpgrp(STDIN_FILENO, getpgid(pid_novi));
 		execvp(naredba[0], naredba);
 		exit(1);
@@ -44,13 +74,26 @@ pid_t pokreni_program(char *naredba[], int u_pozadini)
 	return pid_novi;
 }
 
-int
-main()
+std::string cwd()
+{
+	static const long size = pathconf(".", _PC_PATH_MAX);
+	static char *cwd = nullptr;
+	static char *buf = (char *)malloc((size_t)size);
+	if (buf != nullptr)
+	{
+		cwd = getcwd(buf, (size_t)size);
+	}
+	if (cwd)
+		return cwd;
+	else
+		return "";
+}
+
+int main()
 {
 	struct sigaction act;
 	size_t vel_buf = 128;
 	char buffer[vel_buf];
-	std::list<Proces> still_running;
 	pid_t pid_novi;
 
 	sigemptyset(&act.sa_mask);
@@ -59,39 +102,54 @@ main()
 	sigaction(SIGCHLD, &act, NULL);
 	act.sa_handler = SIG_IGN;
 	sigaction(SIGTTOU, &act, NULL);
-	
+
 	struct termios shell_term_settings;
 	tcgetattr(STDIN_FILENO, &shell_term_settings);
 
+	tcsetpgrp(STDIN_FILENO, getpgid(0));
 
 	do
 	{
 		if (fgets(buffer, vel_buf, stdin) != NULL)
 		{
+			int background = 0;
 			char *argv[MAXARGS];
 			int argc = 0;
 			argv[argc] = strtok(buffer, " \t\n");
 			while (argv[argc] != NULL)
 			{
-				argc++;
-				argv[argc] = strtok(NULL, " \t\n");
-			}
 
-			if (strcmp(argv[0], "cd") == 0 && argc == 2)
-			{
-				static const long size = pathconf(".",_PC_PATH_MAX);
-				static char* cwd = nullptr;
-				static char* buf = (char*)malloc((size_t)size);
-				chdir(argv[1]);
-				if(buf != nullptr)
+				char *value = strtok(NULL, " \t\n");
+				;
+				if (value != NULL)
 				{
-					cwd = getcwd(buf, (size_t)size);
+
+					std::string s_value(value);
+					if (s_value == "&")
+					{
+						std::cout << "Will run in bg" << std::endl;
+						background = 1;
+						continue;
+					}
 				}
-				if(cwd)
-					std::cout << "Changed to: " << cwd << std::endl;
+				argc++;
+				argv[argc] = value;
 			}
 
-			else if (strcmp(argv[0], "ps") == 0 && argc == 1)
+			std::cout << argc << std::endl;
+
+			if (strcmp(argv[0], "pwd") == 0)
+			{
+				std::cout << cwd() << std::endl;
+			}
+
+			else if (strcmp(argv[0], "cd") == 0)
+			{
+				chdir(argv[1]);
+				std::cout << "Changed to " << cwd() << std::endl;
+			}
+
+			else if (strcmp(argv[0], "ps") == 0)
 			{
 				std::cout << "PID\time" << std::endl;
 				for (auto &proces : still_running)
@@ -100,36 +158,43 @@ main()
 				}
 			}
 
-			else if (strcmp(argv[0], "kill") == 0 && argc == 2)
+			else if (strcmp(argv[0], "kill") == 0)
 			{
-				std::cout << "killing " << argv[1] << std::endl;
+				pid_t pid = std::stoi(argv[1]);
+				if (is_running(pid))
+					kill(pid, std::stoi(argv[2]));
+				else std::cout << "Can't kill that." << std::endl;
 			}
-			else 
+			else
 			{
-				pid_novi = pokreni_program(argv, 0);
+				pid_novi = pokreni_program(argv, background);
 				Proces running = Proces(pid_novi, argv[0]);
 				still_running.push_back(running);
 				pid_t pid_zavrsio;
-				do
+				if (!background)
 				{
-					pid_zavrsio = waitpid(pid_novi, NULL, 0);
-					if(pid_zavrsio > 0)
+					do
 					{
-						if(kill(pid_novi, 0) == -1)
+						pid_zavrsio = waitpid(pid_novi, NULL, 0);
+						if (pid_zavrsio > 0)
 						{
-							tcsetpgrp(STDIN_FILENO, getpgid(0));
-							tcsetattr(STDIN_FILENO, 0, &shell_term_settings);	
+							if (kill(pid_novi, 0) == -1)
+							{
+								tcsetpgrp(STDIN_FILENO, getpgid(0));
+								tcsetattr(STDIN_FILENO, 0, &shell_term_settings);
+								ukloni_proces_iz_running(pid_zavrsio);
+							}
+							else
+							{
+								pid_novi = (pid_t)0;
+							}
 						}
-						else 
+						else
 						{
-							pid_novi = (pid_t)0;
+							break;
 						}
-					}
-					else
-					{
-						break;
-					}
-				} while(pid_zavrsio <= 0);
+					} while (pid_zavrsio <= 0);
+				}
 			}
 		}
 	} while (strncmp(buffer, "exit", 4) != 0);
